This is a historical rebuild of GrogLibs, built from a mercurial conversion of what was originally in SVN (and possible perforce before that).
Much of the history was mangled along the way because the importers and
converters often interpreted moves as delete + add new file.  I shall
attempt to fix that and keep the tools, libs, and tests separated.

This might be tricksy as bits of code started out in the tools or tests,
and migrated out to the libs or viceversa.