Fixed two major bugs with lighting.  In one case, I was using the z element of the texturing vectors when it was supposed to be the offset
The other bad spot, I was using the face normal instead of the texture normal in a spot

Changed lightmap scale to 4, probably will need to adjust that a bit as it makes huge lightmaps

Found that some maps have an ambient light value buried in the worldspawn entity.  I have chosen to ignore this.

Did an #if on xbox for the openTitleFile stuff.