massive changes to collada to load in skinned meshes from max.
Currently it works but doesn't save.  Since I can't save out what I build, I'll have to abandon this whole method.

Checking this in to save a working copy of this stuff with all the dynamic type stuff.  I can't use any of that on the xbox though so I'll have to come up with another way.