This is a solid (I think) implementation of genesis up to the save point.
It is extremely ugly, but I found that if I refactor as I go, I break things.  So It is fairly close to the original.

Once light and vis are working and tested on a variety of maps, I'll do a refactor pass.  Gotta remember to time it before and after