XNA 4 conversion
In the original checkin notes here, there was a large, angry and very NSFW rant about the new framework and visual studio 2010.

To summarise, I found build configuration difficult to locate, was angry that build order couldn't be adjusted, and noticed only the last project to build gets shared content.