Don't mark split faces as SIDE_NODE in subtract, this causes major goblinry
Various dump functionality that will be removed, for debuggerizing

Found some problems with matching portals to original brushes.  Sometimes the portal is inverted, but normally the plane number search finds it anyway.  But if it is a non axial plane, it can sometimes take an entirely different plane.  I fixed it by making it check the inverted side as well, but might need to think more about how non axial planes are shared.

Removed a bunch of gacks (debug break spots)

test code in plane sharing stuff