Blasted radiosity, and made a surface light generator that is very slow and needs some work
Increased max portals in vis, eels blew past the end of the array

Added some comments here and there in the libs

Add switchable lights, change mRGBLData to a dictionary so it doesn't waste tons of space

Fix bug with normalizing the color in the light code with a vector normalize instead of a color normalize.  This was causing all maps to be extra dark.

Zone code to handle triggers

Few new entity helpers in ZoneEntity

Switchable light stuff in IndoorMesh and the lightmap Shaders/LightMap.fx

ColorNormalize in RenderHelper

Fix player box size to be like quake1 in BSPTest

BSPTest shows how to properly handle switchable lights