Was incorrectly computing bounds from the bogus polys taken from the map file (where the sides are not clipped yet)
Made a slow but quick and easy face maker to get a valid bounds when needed for model origins and the like

Have some debug code in MoveBox to help with pushing collisions towards the front side.  Will leave it in a bit longer, have breakpoints on the odd cases.

Made my own plane transform function because the xna one is superbroken.  Mine is likely quite slow

Made a generic function to generate 3 points on a plane

Collisions vs a rotated bmodel working well, needs more testing