New Ludum Dare project Minnie & Mal
Disable vis check for sunlight raycast, need to revisit that when head is clear.

Fix odd assert case in a raycast routine for an exactly plane on case.

Turned off the celshading for sky by default, added a technique for that.

Fix sharedform reference on celtweakform in ColladaConvert

Tools are aiming at new project so lots of csproj changes