Need to normalize the cross products for lookat generation stuff!
Warn on wacky raycast behaviours in the sunlight phase

Did a long debugging session on the dark splotches on the walls in my lighting.  Discovered that sometimes raycasts hit right on an edge and could be considered on either side of a node boundary.  This was making the "are you in this face" test fail by just a little bit.

To fix it I reduced the raycast frontside / backside compares to be >0 and allow a little slop in the face pick.

Need to propogate this fix over to zone to fix the occasional errors in the lighthelper

Few more handy calls in Mathery