Clear model stuff when change map in basicmodelhelper
NukeAll in ParticleBoss

Added ApplyForce to Mobile

Attempting to resolve the situation where a caster wants to go to a nearby friend, but the friend is on the other side of a locked door.  Need to implement cutting off PathConnections where doors are if they are locked

A fullbright skin shader to make the eyes glow more reliably

Also MobileAI was not checked in, that could have been disastrous

Now mark path connections to and from locked doors as impassable, and set them to passable when the door is unlocked

MobileAI keeps track of failed pathfinds, adding them to a list of positions and mobs.  Future requests fail if AI attempts to path to them.

Clear these lists whenever a door is unlocked.

DropToGround now has a bool for doing pure world or vs model

Internalized what I could of pathlib

Remove assert when path can't be found

SetPathConnectionPassable finds the node near the location and sets it to pass or not

Only do lighthelper updates if the enemies are in the pvs

RayCollideMobiles also returns a world position hit if it is closer than a mobile impact