Pathfinding now uses a callback to determine if nodes are valid.  For the bsp stuff this calls into the sphere collider.
TryMoveTo doesn't need to check mbOnGround

Replaced TryStandingSpot with TrySphere as the old intersect leaf sides code does not detect if you start in solid.  Really need to replace that.

AI FindSpotAroundPosition now uses the sphere collider stuff to determine if the spot is ok.  This works waaaaaaaaaaay better than the standing spot checker, I've just spotted, it needs to make sure the plane is ground

Detect start in solid for sphere collider, still need to do the others

Penalize via hscore nodes that are too small for a player to get to due to collision.  This lets all the surface area be valid for pathing, but makes sure paths avoid spots they can't fit in.