Make sure standing on ground plane in FindSpotAround
Big mess as I get rid of the old collision routines and replace them with my own.

Don't allow movement under 0.1f in BasicModelHelper.  This helps with getting plane on collisions when pushing against a model.  A bit of a hack but it works.  Will come back later and handle the plane on cases.

ThirdPersonOrient will be broken for a bit

Drop MaxMoveBoxIterations down to 16