Rename intensity to shadowatten
Change directional shadows to use the cubemap style distance code

Change the anim load bone check to check that the new skel has all the old bones, instead of making sure there are no new bones.  This saves a lot of headache if the new anim file has some extra stuff in it.

Lots of adjustments to the light matrix maker for directional.  Stuff was kind of backwards and needed a negative here and there.  Just fiddled with it till it worked