Moved away from the xna style content in a project approach to just aiming all tools and games at a single content folder.  Within that folder is an svn supported shader library, instead of having that in libs (thought it is still there)
The solution now links to the gamey folder for the shaders, so they are still there within the solution for easy access.

Lots of work towards freeing up all resources allocated by the game.  Still a few missing.

Shader compilation now stores timestamps for all headers.  If the files on disk are newer, it recompiles everything since I can't really tell which fx files depend on which fxh.

Textures now read from disk as cpu readable, and have a pass done that premultiplys the alpha and converts the texture to linear space.  This is then copied into a final texture with the original being released.

PostProcess now stores depth textures in the same place as target textures

Work has begun on a system to specify a sort priority for particle systems near to each other.  Haven't yet worked out the details.

Removed a redundant entity keyvalue from GetEntityFields() in Emitter

Use a DeviceDebug in GraphicsDevice to help track down resources