Set release mode path to . so I can give out releases in a zip
Mark core and vis as ok for unsafe in release mode

Fix some stuff related to running out of lightmap atlas space during a zoning

Fix LoadAllMeshes() to work with a . path

Make release mode graphicsdevice not create a debug device.  If you give out a build to someone that doesn't have the sdk installed, it will crash mysteriously

Fix up shader parameter ignores and hides with all of the recent changes to the shaders and renames and such.  Also the bsp auto magical one was using the wrong lists, causing some stuff to be missed