Changed a thing in Jump to do the push velocity thing if the modelOn is >0 instead of not zero.  -1 is for midair.
Removed a duplicate chunk of code in MoveWithoutCollision(), no idea if it was really needed but I think I'd have left a comment if so

Set modelOn to 0 if standing on the world model

Fixed a nasty bug in stair movement that caused stepping on models to use the returned plane from the previous world collision for reflect.  This was causing the resulting movement point to be inside the model.