Include a device enum on the input action so games can scale by user preferences on mouse turn speed or key turn speed or analog or whatever.
Took the mouse multiplier out of input, leaving it to the games for that stuff

ClampInputTimes() in input will chop key held times to a time value specified.  This is useful because the timestamp stuff in the events keeps on going while sitting at a breakpoint, or if the game is running slow and clamps updates to a fixed amount.

Ganked all speedy scaley stuff from PlayerSteering, as well as right click to turn.  That stuff is up to the apps now.