Couple asserts looking for overly large geometry chunks.  Still trying to use uint16 indexes.
DistToNearestEdge() in ConvexPoly.  This provides a metric for rejecting nodes, though I don't think I'm using it right now.  It is not a good metric if checking distances in a poly that is coplanar to the surroundings.

Edge distance to point function.  No idea if this is bulletproof.  I looked at various mathy formulae online and they were absolutely incomprehensible.  So I just create a plane and do an ordinary plane distance check.

When building a grid of path points, added 2 additional checks, and another that is commented out.  First check for points closer than half a gridsize away and reject them, then call the IsValid delegate to look for stuff in the wall or stuck in the ground.  IsValid will attempt to fix the position in the vertical only.  If the point is in the wall it gets rejected.  The other method was distance from edges, but isn't really a good check so is commented out.

For finding connectivity, simply try every move first before getting all fancy with edges.

Commented on a case that hit the e1 & e2 null case