Regular move routine MoveBox can now take a downward step to follow a sloping ramp or step downward.
In mobile move, only step down if starting on the ground and the move along ground flag is set

MoveAlongGroundOnly() for pathing, will probably nuke it

Push sets footing

Trigger wait renamed to Delay and a double is now used to accumulate the time

Set valid foot distance to 1, before it was .4.  Not too sure about that

Movealong now returns true if it adjusted the movement vector

Got rid of zoom and scrollspeed stuff in playersteering

PathGraph skips impassible connections when returning them for draw

When finding connections, only allow upward movement in a stairstep amount, this prevents pathing leaping up onto things.  Jump pathing needs to be a separate step.

double tryparse

Can set nodes to blocked via picking in TestPathing

In CanReach(), return if point a is not on ground.  Not too sure about this one