I was assuming objects were textures in EffectVariableValue, but there are samplers etc... Was causing a really hard to find leak.
Better cleanup when switching techniques in materials.

Use cached type var when checking for texture type in Material (was leaking)

Very difficult to find leak with getting the EffectShaderDescription in layout construction.  Need to split out the EffectShaderVariable and dispose it when done.

Cache technique and pass in postprocess.  All this caching may bite me if there's a device lost condition.  I used to test that via ctrl-alt-del but it doesn't seem to do it anymore.

Clear bindinglist from previous selection when selection changes in material form