Narrowing down the search for move stage entities in BasicModelHelper.cs... An old quake map had a long func_train thing that blew up the stack, might have been a circular linkage.  Should have been done that way anyway.
Leak fix on the mDynLights variable

Lazy inits on pass and technique in post needed a lazy dispose in case they don't ever get allocated (might not be needed)

Since the shadow stage draws in between solids and alphas in IndoorMesh, a callback was needed to let the program reset the rendertargets so the alpha draws don't draw into the shadow targets.  It was never really needed if you had a complicated postprocess thing going but for no post it is needed.