Going back to per plane alpha sorting.  It isn't perfect but per-face was just generating too many draw calls.  I left per-face as an option but it isn't tied to GUI yet.
Alpha draws piled up on a plane can be marked as planar sort.  This happens if they are above a threshold of area which is computed in mapgrinder.  Might be a few material types I still need to do this for, but want to test it heavily first.

Fix a typo that was causing garbage draws from pulling verts out of the wrong arrays.  mLMAAnimVerts instead of mLMAnimVerts... oops.

Alpha sorting looks for nearest planar sort node (if there is one) by plane distance.  It sorts all scene alphas in front or behind, then sorts those buckets and draws them in order depending on the eye position with the big planar node in between.