Simple function to grab the world contents a mobile is in
Changed action mappings to allow more than one map per key combo.  I got a really strong dejavu when doing this and the feeling that this is a very bad idea, but it seems to work fine.

A swim mode of a sort in playersteering.  Acts like fly.

Only apply sprint in playersteering if it makes sense (biped on ground)