A very generic entity library that needs a lot of work
A translator for quake and grog maps to convert entities to the new system.  Barely does anything yet

Grog specific components, so far brush model movers.

Still need lights and pickups and triggers and all sorts of stuff, but things should be better when it is done.  Will probably change alot