Changed the raycollision delegate in CoreDelegates to return the impact point.  The threaded lighting stuff needed this.
Return model impact point when raycollision hits a model.  I'm not entirely sure this is right as the impact point will be in model space.  Might have to correct this in the future.

TextureAxisFromPlane just takes a plane normal now instead of a GBSPPlane so more stuff can use it

As part of resurrecting the old XNA tool LightExplorer, I found the source of some of the lightmap "notching" where a small section would have a unexplained notch in the lighting.  It was the adjust-ray-point-towards-center code.  I've taken that stuff out for now, but there are still errors in baked lighting I'm trying to track down.  My best guess at this point is that there's an incorrect scale with the grid size.  8 seems to work best at the moment.

Set primitive topology in spots that need it

New function to add raw textures to StuffKeeper, used in the light explorer to make little lightmap textures for viewing in 2D

Had some crossed wires in the shader types.  At some point I had read where half types aren't really used on the newer shader models so I think I went in and changed them all to float2, float4 etc.  This caused the automagical input generators to upsize the data they were expecting, as most of my older stuff especially crams stuff like normals into half3.  Thank the gods for RenderDoc.  I'd have never figured out what was going on without it.

Pushed a debug draw module out to SharedForms.  Can't do a proper move as the original file is in the tools repo

StripPath() in fileutil for getting just-the-filename