ScreenUI gumps have some extra parameters for a keyed style gump element, for like rounded meters
Vertex size went up a little for the second set of 16 bit UVs

Store material name in the gump data, so user doesn't need to name the mat a certain way

Spew shader compile warnings into debug text

New gump shaders and techniques in 2d.fx.  I actually struggled with this.  Wanted to use the rounded outline as a key but precompiled alpha screws that up.  Had to just write it as a hard mask.

VPos2Tex04 type