Moved GetCorrectedAngles public as the new gameplay entity stuff needs it
Pickups can be static on the floor, spin, or just have one mesh part spin

Light without a strength defaults to 200 in quake

Moved MakePickUpEnt to the top of the file because VS code's syntax highlighting is flipping out on that file for some reason

SizeY property on ConvexVolume for adjusting raycasts and such

MeshLighting remove that dodgy access of Pickup

Pickup just provides a yaw, doesn't try to set any mats

PosOrient component to handle common position orientation stuffs

Handle setting individual part matrix.  This really should be relative to the parent but right now it is worldspace

New Trilight shader with texture and emissive

Gank BasicModelHelper, LightHelper, QuakeModelHelper, StaticHelper, and TriggerHelper as all that code is now down in entities.  Is it all working?  Probably not.

Future checkins will be attempting new shadowing sorted by light.  The world shadow draw stuff will try to use a batch of 8 lights at once, but I'm not too sure any of this is going to work.  Can you set 8 cubemaps at once for a pass?  If not I've been thinking about doing paraboloid maps for point lights.  That would be 4 lights at once.  Anyway want to have shadowing from every light on every object.