Added some default values to materials
Added a NoBlending blendstate

Was having trouble getting a loaded collada static to draw.  The problem was structure packing.  Going to have to go in and do LayoutKind.Explicit on every vertex structure.  Going to be a pain.  Might just delete most of them.

More work on the material form.  Can pick vertex and pixel shader entrypoints for the material.  Colors can be picked but aren't assigning to the material yet.  Some basic texture pick and display in the works.