? check for null on bone copy
Sometimes need to check the material to see if it needs to be a bspish or characterish material and allocate the substructs and such

Didn't think I would bother with shader reflection this time, but it turned out to be really easy to do.  So made a reflection based input layout maker.

Store a ref to the layout in MeshMaterial.  Should probably release that

Right now I sort of create or get the layout when a material is assigned, but I was thinking at the time that it depended on the geometry involved but it really only depends on the shader used.  Should probably change that.

StructLayout on VPosNormBoneTex0 as I'm messing around with old character test data.  Still looks totally garbled and goblinized.

Change the output form to use the newline stuff that is supposed to be cross plat maybe?  Either way fixed the \n not doing a real return.  Think windows needs \r\n or something.