Taking those little shader structs in material public so I can make a GUI for them.
Function for returning a bitmap of a particular texture to show on a gui or whatever.  The textures are not readable and premultiplied and gammad up so I just reload off the disk.

Texture picker dropdown sort of working.  The display doesn't do some textures very well.  Might need to resize a bit.

Some of the basic functionality of MeshMat is working.  Some elements don't copy their values to the shader until something else does, need to fix that.

Lots of changes to the forms designer so have been using real visual studio for this.  It is awful to start but I kept chipping away at the settings.

Something is odd with the renderererererer.  Time for some renderdoc.