Was only linearizing and premultiplying every 4 pixels on textures, oops.  Manifested as stripes.
Coordinate system stuff fixed the backwardsy light directions I had before.  Fixed trilight computation but specularity is still goofy.  Something is off with the view vector.

Coordinates caused texturing to be effectively flipped vertical.  This isn't a -V but more like a 1 - V.  Fixed it in a pixel shader I've been using but I think there's probably a better place to fix it in a more universal way.  Maybe in the mesh conversion from collada?