Was messing around with the math of the specular functions, thinking there was something off.  I had already ditched the fill light parameter, and it looks better without it.
I implemented schlick's fresnel from math off wiki and then realized it was already doing that in the fresnel stuff.  The only thing that was off was the angles light and the halfvec.  Angles look about right now.  I'm not super picky about the math as I'm usually doing something cartoony anyway.