Started trying to figure out exactly what my stuff's coordinate system was and got very very confused.  Seems like before when I was using the effect framework, camera positions were negated.  After a bit of memory refreshing reading I understand that now.
My player steering stuff moves a matrix around like it is an object matrix, but to use it as a view it needs inverting.  But the math of it can be simplified so that positions and axis vectors can just be negated.

So in the axis vector properties I negate the result, and in the view matrix creation I negate the incoming position.