Changed the way shader compiling works.  Before I was... doing something with headers or something... Not really sure how it worked before to be honest.
Now I check if there's a source folder and if not load all the precompiled I can.

If there is source I check any stored dates of last compilation against the current date of the file.  Any newer get compiled.  Remaining stuff is loaded precompiled.  Seems to work!

If any headers changed I just build everything because I don't really want to try to figure out which source files include which headers.  Doesn't take that long and doesn't happen all that often.

Made a little 1-v change to the emissive shader just to see if the shader compile would fire for that file.  It did!