Some convenience functions to grab DC and CBKeeper so I don't have to pass them around constantly
Function to set a material's shaders and layout instead of pulling them into the mesh draw stuff and doing it there

Got rid of pointers to materialish things in meshes.  These can get out of sync or deleted or reloaded or whatever.  Better to just store names and such and pass the material library in to functions that need it.

Got rid of the trilight setting stuff in meshes.  That should be a materialish thing

Halfs are now supported in file IO, so my character save was writing out halfs without me realizing.  Fix the read to read halfs instead of floats.

Also fixed a int/uint problem with the read/write of color values.

Might end up changing bone indexes to be byte4 to save a little gpu mem.