Sort of halfway started on expanding the skeleton editor to allow for choosing between box/sphere/capsule for each bone with adjusting the size and maybe position if I end up needing it.
Fix clicking on texture1 selecting in the dropdown texture0's texture

Quaternion XYZ() extension when you just want the orientation vector

A bounding capsule that will eventually work in a similar way to box and sphere.  Right now I've written a worldspace raycast to oriented capsule that is totally untested.  I think for most bones this will be the best collision shape.