Frogot to set world matrix in material apply
Check for duplicate on adding an anim to library

Big confusion and mixup on transforms in the whole instance / archetype stuff.  Had this wierd structure I barely remember writing that stored one, like an object space transform, then another in mesh, then another in the top level which would be character or static.

Still need the top level one for moving things about (or placement for statics) but the redunant materialish one was removed, and object transform kept in the mesh part.  This requires passing the top level transform down into the draws.  Should probably reftype that.

Frogot to set shaders in the primitve draws.  It sometimes worked anyway, which is odd.

Fixed capsules to point along the Z axis as the actual bounds do.

The view of the capsules for collision editing is working now, but only in meter units.  I need to have a think about how to properly handle scales.  Right now all animation takes place in the scalespace of the art program, which is in blender's case meters.

When using grog / valve etc units I think a scale is applied to the root bind shape matrix.  This causes goofiness with collision data though.  The collision shapes right now are guessed at by box bounding the influenced verts per bone.  This will be in the art program's scale.

Because the influenced verts are sort of piled up near the origin, the resulting bound doesn't need the bind shape matrix to work.  But if there is a unit scale in there, the sizes and translations will be super off.

Maybe the answer is to involve the bind shape matrixesxeexxeexexseeseseses during bound calculation.

May end up removing box bounds as an option since I don't keep inverted bones around, and can't really think of a way to do oriented boxes without them.