My capsule / ray collide assumed the shortest distance between two lines function was actually for two line segments.  It isn't though, it is for infinite lines.  So alot more checks have to be done to make sure the short segment is in the cylinder section, or intersecting with the endcap spheres.
There was also a bug with the capsule end sphere center using length - radius.  The length is the distance between the sphere centerpoints.  Full length of the capsule is length + radius * 2

Lots of fudgery with transforms.  I was thinking a capsule could directly use the bone transforms without any goblinry, but something is off there and maybe my mental picture of it is off.  Using non bind pose bones from the animated skeleton is giving me the correct position and size.

With all this extra math I'm not super sure this is faster than doing things in bonespace with inverted mats and all.  I shall profile when it is all working properly.

Made a Forward vector extension on Matrix4x4