Ok capsules are perilous.  When I embarked on this endeavour I had the idea that detection would be easy and not needing inverted bones would be a big win.
Hit detection IS easy to be sure, but finding the impact point is a giant pain.  There are so many corner cases.

After spending several days on this I began to cheat and look at other people's solutions.  My own math was missing a square in a term here and there for the forward impact point, causing steep angled collisions to sink into the capsule.

The one I have hooked up now covers all the special cases, and it is LONG.  I've left other routines and failed experiments in there to remind me what a struggle this was.  Next commit I will get rid of the extra junk and hook sphere and box collisions back up.  I also need to give the ray an optional radius.