Sphere and box bone raycast now returning good hits and normals
The box one has a matrix invert in it.  I'm kind of curious what a raycast costs with all boxes vs all spheres vs all capsules.  I might just profile it.

MatrixFromDirection needed a normalize for the first side vector generated.  Near Y directions were making a squishy matrix.

BoxNormalAtPoint returns an axial normal best fit for the passed in point.  I'm using it for surface impacts but it works for whatever.

Got rid of one of the sphere intersects I was experimenting with