Doc notes on instance / archetype stuff.  Not really super happy with the way that all turned out and halfway tempted to rewrite it all.
More work on bounds.  Trying to be consistent with where they are stored, how the initial guess at a bounds is computed etc.  Still alot of work to be done but I wanted to commit this and work on it under the tools repo.  The rough overall bounds need to be editable.

Will probably be allowing ColladaConvert to have a gui to pick which rough bound is better for that particular asset.  It will be saved and used in raycasts as a quick reject.

MeshPartStuff is really annoying and I'm thinking of removing it.  Stuff like raycasts you sort of want in the instance code because that's where the transform is, but then the bound data is stored (in local space) in the arch.  It's all cumbersome and confusing even to me that wrote it.

There was a bit of mixing of the idea of "generate" or "compute" bound and "get" bound.  Will eventually make all of this very clear.

StaticMesh has the new style of ray collide.  Initial quick check based on saved bound choice, then return closest impact on submesh.  Will probably need variants for ones that want every impact, and ones that want the submesh hit returned.