Now have 2 symlinks going into my game directory.  The GrogTools -> GrogLibs -> ShaderLib points to a NTFS symlink at ShadersWin64.  On the linux side Shaders goes to the GrogLibsC ShaderLib.
They might be byte identical but there's some differences in dir structure I got tired of dealing with.

Changed the wic decoder stuff.  CreateDecoderFromFileName crashed down in something that no debugger would step into.  I haven't seen anything like that in ages.  Almost made me want to dig up a copy of softice.  Anyway I worked around it by calling CreateDecoderFromStream() which is pretty much the exact same thing sans crash.

SaveToFile() in character has a frequent failure point.  If the user hasn't set up the bounds yet save would quietly fail.  Need a better way to spam output.

Added pointy arrows onto the ends of the positive axis in commonPrims.  This was to help me wrap my mind around the strange coordinates I was seeing in RenderDoc.  Turned out to be RenderDoc was flipping the x axis.  There's a setting for left / right hand.

Keyframe has alot of experimention code.  When I saw that my left foot was attached to the right bone in blender I went crazy for a day trying to fix it.  Eventually I realized (by looking at my own libdocs.txt) that my coordinate system is the same as Blenders just rotated a bit.  I had it in my noggin that I was using Left handed.  So I was doing alot of goofy experimentation, creating all sorts of Eldritch horrors and abominations.  I want to check it in as is to leave a record for the future in case I forget my own coordinates.  But next commit I am blasting all that conversion stuff.

Fix a bug with world space positions in WNormWPosVS.  I need to have a look at all my shaders that pass a world position in to the pixel stage.  I was doing basically a 3x3 rotation by dot product.  This like calling TransformNormal() in the directX math stuff.  This explains a lot of strange lighting and specular bugs I've seen over the years.

Had a null check in the wrong spot in MaterialForm

Many fixes for Rect going away.  Not sure where it was coming from before but started using Vortice's RawRect.  This is all jumbled in with the changes to a bunch of stuff becoming immutable (mostly boundingboxes).

Device creation was failing on this fresh Win10 because I was trying to create a debug device and didn't have the needed stuff installed.  I handle that now falling back to a regular device.