Label 4 sized texcoords with a 4, was needed in the bsp stuff
No need for reversed winding order

Big giant confusing mess with array types in cbkeeper.  I hope it works.  Will test soon.  It is for dynamic lights and the quakeish light styles.

Input layout generation needed a couple tweaks to watch for 4 component texcoords and full float texcoords that are kind of needed in bsp geoms