Some changes to meshbase, working on ccnogui.  When I switched to a left handed coordinate system awhile back, I never fixed characters.
Brought back some old code to do left handed conversions to skeletons.
For now I'm leaving the character data in right handed and just changing
the root nodes to flip things to left.  This means that collision shapes
and the mesh itself are reversed along X.  I'll fix that shortly.

Warn if AddAnim already has an anim with the same name.  This cost me
alot of time.